Remove Go deployment path and clean up TS/Go naming
Remove initiateGoDeployment from homeView — missing credential or config
is now an error instead of a fallback to the Go backend. Inline
initiateTsDeployment into initiateDeployment since only one path remains.

Delete dead modules: deployProgress.ts (Go SSE progress bridge) and
publishFailures.ts (Go immediate error handler).

Rename runTsDeployWithProgress → runDeployWithProgress,
TsDeployProgressOptions → DeployProgressOptions, and
tsDeployProgress.ts → deployProgress.ts to drop the Go/TS distinction.
